Retrieve comments deserialization fix
