+++ service/src/main/java/com/simplevote/api/comments/objects/MessageObject.java
@@ -1,0 +1,23 @@
+package com.simplevote.api.comments.objects;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class MessageObject {
+    private String message;
+
+    public MessageObject() {
+    }
+
+    public MessageObject(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
